Arreglos MySQL

+++ contratos-laravel/database/migrations/2021_04_22_214425_create_ciudads_table.php
@@ -15,7 +15,7 @@ class CreateCiudadsTable extends Migration
     {
         Schema::create('ciudads', function (Blueprint $table) {
             $table->id('ID_ciudad');
-            $table->integer('ID_region')->unsigned();
+            $table->bigInteger('ID_region')->unsigned();
             $table->foreign('ID_region')->references('ID_region')->on('regions');
             $table->String('Nombre_c');
         });
